requirements.txt
PyYAML==5.4.1
eth-brownie==1.17.1
pytest==6.2.5